Rename file : logging_blocks -> prebuilt_logs
